backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.10.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.27.1
praw==7.8.1
google-generativeai==0.8.3
apscheduler==3.10.4

sendgrid==6.11.0
pillow==11.0.0
python-multipart==0.0.17
httpx==0.28.1
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-telegram-bot==21.8
tweepy==4.14.0
instagrapi==2.1.2
facebook-sdk==3.1.0
linkedin-api==2.2.0
redis==5.2.0
celery==5.4.0
pytest==8.3.4
pytest-asyncio==0.24.0
